moved query validator test
